LayoutTests/fast/forms/ios/select-multiple-options-after-optgroup-expected.txt

This test verifies that choosing multiple options which have preceding optgroup siblings works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectValues is ""
PASS selectValues is "B"
PASS selectValues is "B,2"
PASS selectValues is "B,2,D"
PASS selectValues is "B,2,D,E"
PASS successfullyParsed is true

TEST COMPLETE
